# [FabricClient] Service notification perf comparison